Add custom data grid cell rendering
