+++ plugins/main/public/components/common/data_grid/index.ts
@@ -1,0 +1,2 @@
+export * from './data_grid_service';
+export * from './use_data_grid';
